More Gtk/XmHTML work: now a number of features can be configured on the widget, look at the gazillions of options in gtk-xmhtml.h for more information.  This is probably not complete, I have to read over the XmHTML/Motif docs for figuring which other bits I am missing.
Known bug: the parser is not fixing documents without <BODY> tags,
so this means that the program is crashing if you pass a document
withou the BODY.   I bet this is caused by some mistake of mine in
the handling of the old/new/current widgets as created by Motif.

Miguel.

svn path=/trunk/; revision=64