OK, fixed the non-working images on 16 bpp.  This is done by putting in a my_gdk_image_new() that explicitly takes the data, padding and bytes_per_line parameters for XCreateImage().  The function is a copy of the GDK_IMAGE_NORMAL part of the real gdk_image_new(); it just uses the specified parameters instead of the Gdk hardcoded values.
This is needed because by default gdk_image_new() always creates
images with a 32-bit padding for scanclines.  XmHTML, however, uses
different paddings depending on the image depth.

As a sidenote, does anyone know where to get documentation for XShm?
I am not using GDK_IMAGE_FASTEST because I do not see a way to pass
the padding parameter to the XShm functions.

- Federico

svn path=/trunk/; revision=123