Don't crash if p->section_name == NULL. (*): Use `g_strdup' and `g_free'
1998-02-28  Raja R Harinath  <harinath@cs.umn.edu>

        * libgnome/gnome-config.c (dump_sections): Don't crash if
        p->section_name == NULL.
        (*): Use `g_strdup' and `g_free' throughout.

(This is a better fix than my earlier braindead replacement of NULL with "".
The essential problem it is trying to fix is that `strlen(NULL)' can crash...
and strlen is called in strdup, and *printf for some.  `g_strdup' is used
since it calls `g_malloc', and is hence compatible with the rest of the
allocations, which call `g_{new,malloc}'.

svn path=/trunk/; revision=459