I am writing documentation for the gnome-app-helper module. It is still
I am writing documentation for the gnome-app-helper module.  It is still not
finished, but it is going well.  I would appreciate it if some DocBook wizard
could check gnome-app-helper.sgml for bloopers and such.  So far the most
useful DocBook documentation I have been able to find is the examples in the
reference manual :-(

1998-09-20  Federico Mena Quintero  <federico@nuclecu.unam.mx>

	* gnome-app-helper.sgml: New file with documentation for the
	gnome-app-helper module.  This is a work in progress.  PSGML mode
	for Emacs rocks, btw :-)

	* Makefile.am (EXTRA_DIST): Added gnome-app-helper.sgml to the
	list of files.

svn path=/trunk/; revision=1532