Ok, got a suid wrapper to do wtmp/utmp updates and open the ptys and set
Ok, got a suid wrapper to do wtmp/utmp updates and open the ptys and set
the permissions correctly.  This works on Linux and *might* work correctly
in Solaris, but it really needs people to provide their OS versions for the
gnome-utmp.c file.

I also documented the Zvt API with the embedded doc thing while I was doing
this.

Miguel.

svn path=/trunk/; revision=1906