A tool to generate documentation from C files.
Generates html directly, or via DocBook - using REFENTRY
sections for each function.  Can process many files/functions
or only spit out individual functions.  The code's a bit messy,
but it works fairly robustly.

Supports a superset of the api comment format in devel-docs.
The extra functionality is mostly the ability to define any number of
arbitrary sections.

The html output from sgmltools is _ok_, but a bit 'airy'.  The
postscript was broken when I tried it - though what worked was extremely
airy, and not very readable.

svn path=/trunk/; revision=1924