This is an emacs-lisp file which can aid writing the comment headers for gnome api documentation.
It provides a function "gnome-doc-insert", which will create a properly
formatted skeleton comment in the "gnome-style", including all parameter
and Return Value (if the function isn't a void function) fields, for
the 'current' c function.

By default this is bound to C-x4h (for 'insert header', similar keystrokes
to add-changelog for ease of reference :).

svn path=/trunk/; revision=1932