First steps for a reliable gnome-name-server.
First steps for a reliable gnome-name-server.

Note that this name server only implements the reliability
bits.  The actual code for the name server will be kept in
ORBit/src/service/name

Tomorrow I will make the ORBit name server be available in library
form, to allow me to link this with the existing code.  So, dont
be afraid, there wont be any code duplication.

Miguel.

svn path=/trunk/; revision=2343