No, we will not have oneway:
No, we will not have oneway:

	1. This is an over-optimization;  There is no proof it is
	   benefical, while making the semantics blury.

	2. Current semantics are:

		a. You dont get an exception, things are fine.
		b. You get an exception, things are gone.

Given that oneway does not produce any *measured* benefits and given
Matt's excelent description of oneway semantics (ie, best delivery, with
no guarantee of service) we will kill oneway.

Why do we kill one-way?  Because we do not want "ref" to be "best-efforted",
ie, we can not afford a "ref" to be lost.

So, indeed unref() is a good-efford garbage collection and is not trying to
solve any GC research problems.

But ref() is a you better get it right or you lose.

Do not modify this file without the gnome-libs and bonobo maintainer approval,
That would be me ;-)

Miguel.

svn path=/trunk/; revision=3911