*references.sgml: New file which contains the bibliography that was in ui-guide.sgml. The file nw uses bibliography markup. I also added some books to the bibliography.
*ui-guide.sgml: Added references.sgml to the list of entities
and replaced bibliogrpahy with entity.
Added Martin Wahlen (who does not work on the document any more,
but has contributed earlier by porting it to DocBook) and myself
to the authors section.
Moved the content of about.sgml into an abstract element in the
bookinfo (ui-guide.sgml).
Added a first draft of the planned structure as a comment.

*about.sgml: Removed. The content is now in ui-guide.sgml. The
chapter will eventually be replaced by an introduction.

svn path=/trunk/; revision=4871