*** IMPORTANT ***
This is the _development branch_ of gnome-libs.  It is meant for developers
of gnome-libs only!

Please do not translate this version of gnome-libs, it is a perfect waste of
your time.  There is not a single person who would be able to benefit from your
translations since you must not use this version of gnome-libs yet.  It is meant
for developers of gnome-libs only!

If you want to do any translations, please do them in the stable branch, this is
the version of gnome-libs which everybody is using.

 *** You should be using gnome-libs 1.0 instead. ***

If you are using this source tree from CVS, please run:
	cvs update -r gnome-libs-1-0"
now.


August 7, 2000
Martin Baulig <baulig@suse.de>

svn path=/trunk/; revision=5651